+++ mocks/account/main.js
@@ -7,7 +7,8 @@ module.exports = function(http) {
     var fn = require('../utils/fn'),
         db = require('../utils/db'),
         idFromPath = db.idFromPath,
-        extend = fn.extend;
+        extend = fn.extend,
+        genId = require('../../tasks/resources/helpers').genId;
 
     function objectPath(type, id) {
         return path.resolve(__dirname, './' + type + '/' + id + '.json');
@@ -50,6 +51,50 @@ module.exports = function(http) {
         } catch(e) {
             this.respond(401, 'Not Authorized');
         }
+    });
+
+    http.whenPOST('/api/account/users/confirm/**', function(request) {
+        var id = idFromPath(request.pathname),
+            filePath = objectPath('users', id),
+            token = request.body.token,
+            currentTime = (new Date()).toISOString(),
+            user;
+
+        if (token) {
+            try {
+                user = extend(grunt.file.readJSON(filePath), {
+                    advertiser: "a-282824b8bb40a2",
+                    customer: "cus-71e725f8bf33d5",
+                    created: currentTime,
+                    lastUpdated: currentTime,
+                    org: "o-a6fd7298acb6fa",
+                    status: "active"
+                });
+
+                grunt.file.write(filePath, JSON.stringify(user, null, '    '));
+                this.respond(200, extend(grunt.file.readJSON(filePath), { id: id }));
+            } catch(e) {
+                this.respond(401, 'Not Authorized');
+            }
+        } else {
+            this.respond(403, 'Forbidden');
+        }
+    });
+
+    http.whenPOST('/api/account/users/signup', function(request) {
+        var id = genId('u'),
+            currentTime = (new Date()).toISOString(),
+            user = extend(request.body, {
+                applications: ["e-99263e70058290"],
+                config: {},
+                created: currentTime,
+                lastUpdated: currentTime,
+                status: 'new'
+            });
+
+        grunt.file.write(objectPath('users', id), JSON.stringify(user, null, '    '));
+
+        this.respond(201, extend(user, { id: id }));
     });
 
     /***********************************************************************************************