+++ test/spec/selfie.state.ut.js
@@ -70,8 +70,22 @@ define(['app'], function(appModule) {
                 });
 
                 it('should transition to the login state', function() {
-                    expect(c6State.goTo).toHaveBeenCalledWith('Login', null, null, true);
+                    expect(c6State.goTo).toHaveBeenCalledWith('Selfie:Login', null, null, true);
                 });
+            });
+        });
+
+        describe('afterModel()', function() {
+            it('should go to Selfie:ResendActivation if the user has status of "new"', function() {
+                spyOn(c6State,'goTo');
+
+                selfie.afterModel({status: 'active'});
+
+                expect(c6State.goTo).not.toHaveBeenCalled();
+
+                selfie.afterModel({status: 'new'});
+
+                expect(c6State.goTo).toHaveBeenCalledWith('Selfie:ResendActivation', ['You are a new user']);
             });
         });
 
@@ -84,15 +98,23 @@ define(['app'], function(appModule) {
                 };
 
                 spyOn(c6State, 'goTo');
+            });
 
+            it('should go to the "Selfie:Apps" state', function() {
                 selfie.cModel = user;
                 $rootScope.$apply(function() {
                     selfie.enter();
                 });
+                expect(c6State.goTo).toHaveBeenCalledWith('Selfie:Apps', null, null, true);
             });
 
-            it('should go to the "Selfie:Apps" state', function() {
-                expect(c6State.goTo).toHaveBeenCalledWith('Selfie:Apps', null, null, true);
+            it('should go to the Selfie:ResendActivation if the user has status of "new"', function() {
+                selfie.cModel = user;
+                selfie.cModel.status = 'new';
+                $rootScope.$apply(function() {
+                    selfie.enter();
+                });
+                expect(c6State.goTo).toHaveBeenCalledWith('Selfie:ResendActivation', ['You are a new user']);
             });
         });
     });